Rename compaction package to compacted
