+++ slatedb/compactor_state.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/slatedb/slatedb-go/internal/assert"
 	"github.com/slatedb/slatedb-go/internal/sstable"
-	compaction2 "github.com/slatedb/slatedb-go/slatedb/compaction"
+	compaction2 "github.com/slatedb/slatedb-go/slatedb/compacted"
 	"github.com/slatedb/slatedb-go/slatedb/state"
 
 	"github.com/oklog/ulid/v2"
