+++ slatedb/compactor_state_test.go
@@ -11,7 +11,7 @@ import (
 
 	assert2 "github.com/slatedb/slatedb-go/internal/assert"
 	"github.com/slatedb/slatedb-go/internal/sstable"
-	compaction2 "github.com/slatedb/slatedb-go/slatedb/compaction"
+	compaction2 "github.com/slatedb/slatedb-go/slatedb/compacted"
 	"github.com/slatedb/slatedb-go/slatedb/config"
 	"github.com/slatedb/slatedb-go/slatedb/state"
 	"github.com/slatedb/slatedb-go/slatedb/store"
