+++ slatedb/sortedrun_test.go
@@ -14,7 +14,7 @@ import (
 	"github.com/slatedb/slatedb-go/internal/sstable"
 	"github.com/slatedb/slatedb-go/internal/types"
 	"github.com/slatedb/slatedb-go/slatedb/common"
-	"github.com/slatedb/slatedb-go/slatedb/compaction"
+	"github.com/slatedb/slatedb-go/slatedb/compacted"
 	"github.com/slatedb/slatedb-go/slatedb/store"
 )
 
@@ -24,14 +24,14 @@ func buildSRWithSSTs(
 	tableStore *store.TableStore,
 	keyGen common.OrderedBytesGenerator,
 	valGen common.OrderedBytesGenerator,
-) (compaction.SortedRun, error) {
+) (compacted.SortedRun, error) {
 
 	sstList := make([]sstable.Handle, 0, n)
 	for i := uint64(0); i < n; i++ {
 		writer := tableStore.TableWriter(sstable.NewIDCompacted(ulid.Make()))
 		for j := uint64(0); j < keysPerSST; j++ {
 			if err := writer.Add(keyGen.Next(), mo.Some(valGen.Next())); err != nil {
-				return compaction.SortedRun{}, err
+				return compacted.SortedRun{}, err
 			}
 		}
 
@@ -39,7 +39,7 @@ func buildSRWithSSTs(
 		sstList = append(sstList, *sst)
 	}
 
-	return compaction.SortedRun{ID: 0, SSTList: sstList}, nil
+	return compacted.SortedRun{ID: 0, SSTList: sstList}, nil
 }
 
 func TestOneSstSRIter(t *testing.T) {
@@ -58,8 +58,8 @@ func TestOneSstSRIter(t *testing.T) {
 	sstHandle, err := tableStore.WriteSST(sstable.NewIDCompacted(ulid.Make()), encodedSST)
 	assert.NoError(t, err)
 
-	sr := compaction.SortedRun{ID: 0, SSTList: []sstable.Handle{*sstHandle}}
-	iterator, err := compaction.NewSortedRunIterator(sr, tableStore)
+	sr := compacted.SortedRun{ID: 0, SSTList: []sstable.Handle{*sstHandle}}
+	iterator, err := compacted.NewSortedRunIterator(sr, tableStore)
 	assert.NoError(t, err)
 	assert2.Next(t, iterator, []byte("key1"), []byte("value1"))
 	assert2.Next(t, iterator, []byte("key2"), []byte("value2"))
@@ -94,8 +94,8 @@ func TestManySstSRIter(t *testing.T) {
 	sstHandle2, err := tableStore.WriteSST(sstable.NewIDCompacted(ulid.Make()), encodedSST)
 	require.NoError(t, err)
 
-	sr := compaction.SortedRun{ID: 0, SSTList: []sstable.Handle{*sstHandle, *sstHandle2}}
-	iterator, err := compaction.NewSortedRunIterator(sr, tableStore)
+	sr := compacted.SortedRun{ID: 0, SSTList: []sstable.Handle{*sstHandle, *sstHandle2}}
+	iterator, err := compacted.NewSortedRunIterator(sr, tableStore)
 	assert.NoError(t, err)
 	assert2.Next(t, iterator, []byte("key1"), []byte("value1"))
 	assert2.Next(t, iterator, []byte("key2"), []byte("value2"))
@@ -129,7 +129,7 @@ func TestSRIterFromKey(t *testing.T) {
 		fromKey := testCaseKeyGen.Next()
 		testCaseValGen.Next()
 
-		kvIter, err := compaction.NewSortedRunIteratorFromKey(sr, fromKey, tableStore)
+		kvIter, err := compacted.NewSortedRunIteratorFromKey(sr, fromKey, tableStore)
 		assert.NoError(t, err)
 
 		for j := 0; j < 30-i; j++ {
@@ -158,7 +158,7 @@ func TestSRIterFromKeyLowerThanRange(t *testing.T) {
 	sr, err := buildSRWithSSTs(3, 10, tableStore, keyGen, valGen)
 	require.NoError(t, err)
 
-	kvIter, err := compaction.NewSortedRunIteratorFromKey(sr, []byte("aaaaaaaaaa"), tableStore)
+	kvIter, err := compacted.NewSortedRunIteratorFromKey(sr, []byte("aaaaaaaaaa"), tableStore)
 	assert.NoError(t, err)
 
 	for j := 0; j < 30; j++ {
@@ -184,7 +184,7 @@ func TestSRIterFromKeyHigherThanRange(t *testing.T) {
 	sr, err := buildSRWithSSTs(3, 10, tableStore, keyGen, valGen)
 	require.NoError(t, err)
 
-	kvIter, err := compaction.NewSortedRunIteratorFromKey(sr, []byte("zzzzzzzzzzzzzzzzzzzzzzzzzzzzzz"), tableStore)
+	kvIter, err := compacted.NewSortedRunIteratorFromKey(sr, []byte("zzzzzzzzzzzzzzzzzzzzzzzzzzzzzz"), tableStore)
 	assert.NoError(t, err)
 	next, ok := kvIter.Next(context.Background())
 	assert.False(t, ok)
